exclude logic to service
