+++ images/virtualization-artifact/pkg/controller/vm/internal/validators/sizing_policy_compliency_validator.go
@@ -18,29 +18,28 @@ package validators
 
 import (
 	"context"
-	"errors"
-	"fmt"
-	"strconv"
-	"strings"
 
-	"k8s.io/apimachinery/pkg/api/resource"
-	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 
+	"github.com/deckhouse/virtualization-controller/pkg/controller/service"
 	"github.com/deckhouse/virtualization/api/core/v1alpha2"
 )
 
 type SizingPolicyCompliencyValidator struct {
-	client client.Reader
+	client  client.Reader
+	service service.SizePolicyService
 }
 
 func NewSizingPolicyCompliencyValidator(client client.Reader) *SizingPolicyCompliencyValidator {
-	return &SizingPolicyCompliencyValidator{client: client}
+	return &SizingPolicyCompliencyValidator{
+		client:  client,
+		service: *service.NewSizePolicyService(client),
+	}
 }
 
 func (v *SizingPolicyCompliencyValidator) ValidateCreate(ctx context.Context, vm *v1alpha2.VirtualMachine) (admission.Warnings, error) {
-	err := v.CheckVMCompliedSizePolicy(ctx, vm)
+	err := v.service.CheckVMCompliedSizePolicy(ctx, vm)
 	if err != nil {
 		return nil, err
 	}
@@ -49,183 +48,10 @@ func (v *SizingPolicyCompliencyValidator) ValidateCreate(ctx context.Context, vm
 }
 
 func (v *SizingPolicyCompliencyValidator) ValidateUpdate(ctx context.Context, _, newVM *v1alpha2.VirtualMachine) (admission.Warnings, error) {
-	err := v.CheckVMCompliedSizePolicy(ctx, newVM)
+	err := v.service.CheckVMCompliedSizePolicy(ctx, newVM)
 	if err != nil {
 		return nil, err
 	}
 
 	return nil, nil
-}
-
-func (v *SizingPolicyCompliencyValidator) CheckVMCompliedSizePolicy(ctx context.Context, vm *v1alpha2.VirtualMachine) error {
-	vmclass := &v1alpha2.VirtualMachineClass{}
-	err := v.client.Get(ctx, types.NamespacedName{
-		Name: vm.Spec.VirtualMachineClassName,
-	}, vmclass)
-	if err != nil {
-		return err
-	}
-
-	sizePolicy := getVMSizePolicy(vm, vmclass)
-	if sizePolicy == nil {
-		return fmt.Errorf(
-			"virtual machine %s has no valid size policy in vm class %s",
-			vm.Name, vm.Spec.VirtualMachineClassName,
-		)
-	}
-
-	var errorsArray []error
-
-	errorsArray = append(errorsArray, validateCoreFraction(vm, sizePolicy)...)
-	errorsArray = append(errorsArray, validateVMMemory(vm, sizePolicy)...)
-	errorsArray = append(errorsArray, validatePerCoreMemory(vm, sizePolicy)...)
-
-	if len(errorsArray) > 0 {
-		return fmt.Errorf("errors while size policy validate: %w", errors.Join(errorsArray...))
-	}
-
-	return nil
-}
-
-func getVMSizePolicy(vm *v1alpha2.VirtualMachine, vmclass *v1alpha2.VirtualMachineClass) *v1alpha2.SizingPolicy {
-	for _, sp := range vmclass.Spec.SizingPolicies {
-		if vm.Spec.CPU.Cores >= sp.Cores.Min && vm.Spec.CPU.Cores <= sp.Cores.Max {
-			return sp.DeepCopy()
-		}
-	}
-
-	return nil
-}
-
-func validateCoreFraction(vm *v1alpha2.VirtualMachine, sp *v1alpha2.SizingPolicy) (errorsArray []error) {
-	if sp.CoreFractions == nil {
-		return
-	}
-
-	fractionStr := strings.ReplaceAll(vm.Spec.CPU.CoreFraction, "%", "")
-	fraction, err := strconv.Atoi(fractionStr)
-	if err != nil {
-		errorsArray = append(errorsArray, fmt.Errorf("cpu fraction value is invalid"))
-		return
-	}
-
-	hasInSizePolicyFractions := false
-	for _, spFraction := range sp.CoreFractions {
-		if fraction == int(spFraction) {
-			hasInSizePolicyFractions = true
-		}
-	}
-
-	if !hasInSizePolicyFractions {
-		errorsArray = append(errorsArray, fmt.Errorf("vm core fraction incorrect"))
-	}
-
-	return
-}
-
-func validateVMMemory(vm *v1alpha2.VirtualMachine, sp *v1alpha2.SizingPolicy) (errorsArray []error) {
-	if sp.Memory == nil || sp.Memory.Max.IsZero() {
-		return
-	}
-
-	if vm.Spec.Memory.Size.Cmp(sp.Memory.Min) == -1 {
-		errorsArray = append(errorsArray, fmt.Errorf(
-			"requested VM memory (%s) lesser than valid minimum, available range [%s, %s]",
-			vm.Spec.Memory.Size.String(),
-			sp.Memory.Min.String(),
-			sp.Memory.Max.String(),
-		))
-	}
-
-	if vm.Spec.Memory.Size.Cmp(sp.Memory.Max) == 1 {
-		errorsArray = append(errorsArray, fmt.Errorf(
-			"requested VM memory (%s) greater than valid maximum, available range [%s, %s]",
-			vm.Spec.Memory.Size.String(),
-			sp.Memory.Min.String(),
-			sp.Memory.Max.String(),
-		))
-	}
-
-	if !sp.Memory.Step.IsZero() {
-		err := checkInGrid(vm.Spec.Memory.Size, sp.Memory.Min, sp.Memory.Max, sp.Memory.Step, "VM memory")
-		if err != nil {
-			errorsArray = append(errorsArray, err)
-		}
-	}
-
-	return
-}
-
-func validatePerCoreMemory(vm *v1alpha2.VirtualMachine, sp *v1alpha2.SizingPolicy) (errorsArray []error) {
-	if sp.Memory == nil || sp.Memory.PerCore.Max.IsZero() {
-		return
-	}
-
-	// not have a default dividing :(
-	// dirty, I know
-	// wash your hands after read this
-	vmPerCore := vm.Spec.Memory.Size.Value() / int64(vm.Spec.CPU.Cores)
-	perCoreMemory := resource.NewQuantity(vmPerCore, resource.BinarySI)
-
-	if perCoreMemory.Cmp(sp.Memory.PerCore.Min) == -1 {
-		errorsArray = append(errorsArray, fmt.Errorf(
-			"requested VM per core memory (%s) lesser than valid minimum, available range [%s, %s]",
-			perCoreMemory.String(),
-			sp.Memory.PerCore.Min.String(),
-			sp.Memory.PerCore.Max.String(),
-		))
-	}
-
-	if perCoreMemory.Cmp(sp.Memory.PerCore.Max) == 1 {
-		errorsArray = append(errorsArray, fmt.Errorf(
-			"requested VM per core memory (%s) greater than valid maximum, available range [%s, %s]",
-			perCoreMemory.String(),
-			sp.Memory.PerCore.Min.String(),
-			sp.Memory.PerCore.Max.String(),
-		))
-	}
-
-	if !sp.Memory.Step.IsZero() {
-		err := checkInGrid(*perCoreMemory, sp.Memory.PerCore.Min, sp.Memory.PerCore.Max, sp.Memory.Step, "VM per core memory")
-		if err != nil {
-			errorsArray = append(errorsArray, err)
-		}
-	}
-
-	return
-}
-
-func checkInGrid(value, min, max, step resource.Quantity, source string) (err error) {
-	grid := generateValidGrid(min, max, step)
-
-	for i := 0; i < len(grid)-1; i++ {
-		cmpLeftResult := value.Cmp(grid[i])
-		cmpRightResult := value.Cmp(grid[i+1])
-
-		if cmpLeftResult == 0 || cmpRightResult == 0 {
-			return
-		} else if cmpLeftResult == 1 && cmpRightResult == -1 {
-			err = fmt.Errorf(
-				"requested %s not in available values grid, nearest valid values [%s, %s]",
-				source,
-				grid[i].String(),
-				grid[i+1].String(),
-			)
-			return
-		}
-	}
-
-	return
-}
-
-func generateValidGrid(min, max, step resource.Quantity) []resource.Quantity {
-	var grid []resource.Quantity
-
-	for val := min; val.Cmp(max) == -1; val.Add(step) {
-		grid = append(grid, val)
-	}
-
-	grid = append(grid, max)
-
-	return grid
 }
